Game: updateX split up
seems sensible, so that we can update Ships/Shots separately later
whenever we receive updates from the server. It's not necessary anymore
to update ALL the ships/shots, but we can just do it with a few.

The alternative would have been to do sth like this on the receival of
an dt old packet:
```C++
updateX(-dt);
insert X(newData);
updateX(dt);
```
which would take a little bit of time